pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
    "torch",
    "torchvision",
    "safetensors",
    "accelerate",
    "open_clip_torch",
    "loguru",
    "Pillow",
    "fastapi",
    "pydantic",
    "rtoml",
    "uvicorn",
    "python-multipart",
    "transformers>4.27.1",
    "absl-py>=2.0.0",
]
requires-python = ">=3.9"
